Core rewrite and split into files
Gulp is now being used to concat multiple files in order to allow
splitting the different objects up into different classes. Also
contains a rewrite of the MessageBotCore object.
