+++ .gitignore
@@ -1,0 +1,5 @@
+#Node
+node_modules
+
+#System files
+.DS_Store
