+++ dev/footer.js
@@ -1,0 +1,5 @@
+/*global
+	MessageBot
+*/
+var bot = new MessageBot();
+bot.start();
