+++ package.json
@@ -1,0 +1,25 @@
+{
+  "name": "blockheads-messagebot",
+  "version": "5.0.0",
+  "description": "A bot for the Blockheads Cloud Portal",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Bibliofile/Blockheads-MessageBot.git"
+  },
+  "author": "Bibliophile",
+  "license": "SEE LICENSE IN LICENSE",
+  "bugs": {
+    "url": "https://github.com/Bibliofile/Blockheads-MessageBot/issues"
+  },
+  "homepage": "https://github.com/Bibliofile/Blockheads-MessageBot#readme",
+  "devDependencies": {
+    "babel-preset-es2015": "^6.6.0",
+    "babel-preset-es2015-loose": "^7.0.0",
+    "gulp": "^3.9.1",
+    "gulp-babel": "^6.1.2",
+    "gulp-concat": "^2.6.0",
+    "gulp-jshint": "^2.0.0",
+    "gulp-strip-comments": "^2.4.2",
+    "jshint": "^2.9.1"
+  }
+}
